jajca

+++ apps/backend/migrations/2025-03-27-163647_create_users/up.sql
@@ -1,7 +1,7 @@
 -- Your SQL goes here
 CREATE TABLE users (
     id SERIAL PRIMARY KEY,
-        username VARCHAR(255) NOT NULL UNIQUE,
-        password VARCHAR(255) NOT NULL,
-        created_at INTEGER NOT NULL
+    username VARCHAR(255) NOT NULL UNIQUE,
+    password VARCHAR(255) NOT NULL,
+    created_at BIGINT NOT NULL
 );
